Simplify wallet sync
The last commit introduced `apply_block_relevant` for
BdkWallet. Using this we can greatly simplify the logic
of syncing via FilterIter. Now instead of cloning out the
keychains and tracking changes in a temporary
IndexedTxGraph, we can apply matching blocks directly
to the wallet. Updating the local chain works the same
as before.

To make it work we have to initialize a fresh RPC
client to avoid holding both a mutable and shared
reference to the Coordinator simultaneously, which
is slightly annoying but I have not yet devised a better
solution.